Share close_fds Command function
