+++ libs/utils/Cargo.toml
@@ -20,6 +20,7 @@ tokio-rustls = "0.23"
 tracing = "0.1"
 tracing-subscriber = { version = "0.3", features = ["env-filter", "json"] }
 nix = "0.25"
+close_fds = "0.3.2"
 signal-hook = "0.3.10"
 rand = "0.8.3"
 jsonwebtoken = "8"
