+++ libs/utils/src/fs_ext.rs
@@ -1,5 +1,5 @@
 /// Extensions to `std::fs` types.
-use std::{fs, io, path::Path};
+use std::{fs, io, os::unix::process::CommandExt, path::Path, process::Command};
 
 pub trait PathExt {
     /// Returns an error if `self` is not a directory.
@@ -12,6 +12,40 @@ where
 {
     fn is_empty_dir(&self) -> io::Result<bool> {
         Ok(fs::read_dir(self)?.into_iter().next().is_none())
+    }
+}
+
+///
+/// Command with ability not to give all file descriptors to child process
+///
+pub trait CloseFileDescriptors: CommandExt {
+    ///
+    /// Close file descriptors (other than stdin, stdout, stderr) in child process
+    ///
+    fn close_fds(&mut self) -> &mut Command;
+}
+
+impl<C: CommandExt> CloseFileDescriptors for C {
+    fn close_fds(&mut self) -> &mut Command {
+        unsafe {
+            self.pre_exec(move || {
+                // SAFETY: Code executed inside pre_exec should have async-signal-safety,
+                // which means it should be safe to execute inside a signal handler.
+                // The precise meaning depends on platform. See `man signal-safety`
+                // for the linux definition.
+                //
+                // The set_fds_cloexec_threadsafe function is documented to be
+                // async-signal-safe.
+                //
+                // Aside from this function, the rest of the code is re-entrant and
+                // doesn't make any syscalls. We're just passing constants.
+                //
+                // NOTE: It's easy to indirectly cause a malloc or lock a mutex,
+                // which is not async-signal-safe. Be careful.
+                close_fds::set_fds_cloexec_threadsafe(3, &[]);
+                Ok(())
+            })
+        }
     }
 }
